deployment/backend/requirements.txt
waitress
flask
pandas
numpy
scikit-learn
tensorflow-cpu